Simply CollectionHeader by splitting into two
